Refacto Menu + fermeture lors d'un changement de lien
