+++ src/containers/Menu/MenuMobile.jsx
@@ -1,0 +1,138 @@
+import React, { useState, useEffect } from "react";
+import AppBar from "@material-ui/core/AppBar";
+import { NavLink, useLocation } from "react-router-dom";
+import {
+  Divider,
+  Drawer,
+  Hidden,
+  IconButton,
+  List,
+  ListItem,
+  ListItemIcon,
+  ListItemText,
+  makeStyles,
+  Tab,
+  Toolbar,
+  useTheme
+} from "@material-ui/core";
+import MenuIcon from "@material-ui/icons/Menu";
+import ChevronLeftIcon from "@material-ui/icons/ChevronLeft";
+import ChevronRightIcon from "@material-ui/icons/ChevronRight";
+const drawerWidth = 240;
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    display: "flex"
+  },
+  appBar: {
+    transition: theme.transitions.create(["margin", "width"], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    })
+  },
+  appBarShift: {
+    width: `calc(100% - ${drawerWidth}px)`,
+    marginLeft: drawerWidth,
+    transition: theme.transitions.create(["margin", "width"], {
+      easing: theme.transitions.easing.easeOut,
+      duration: theme.transitions.duration.enteringScreen
+    })
+  },
+  menuButton: {
+    marginRight: theme.spacing(2)
+  },
+  hide: {
+    display: "none"
+  },
+  drawer: {
+    width: drawerWidth,
+    flexShrink: 0
+  },
+  drawerPaper: {
+    width: drawerWidth
+  },
+  drawerHeader: {
+    display: "flex",
+    alignItems: "center",
+    padding: theme.spacing(0, 1),
+    // necessary for content to be below app bar
+    ...theme.mixins.toolbar,
+    justifyContent: "flex-end"
+  },
+  content: {
+    flexGrow: 1,
+    padding: theme.spacing(3),
+    transition: theme.transitions.create("margin", {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    }),
+    marginLeft: -drawerWidth
+  },
+  contentShift: {
+    transition: theme.transitions.create("margin", {
+      easing: theme.transitions.easing.easeOut,
+      duration: theme.transitions.duration.enteringScreen
+    }),
+    marginLeft: 0
+  }
+}));
+
+const MenuMobile = ({ links }) => {
+  const [open, setOpen] = useState(false);
+  const classes = useStyles();
+  const theme = useTheme();
+  const location = useLocation();
+
+  useEffect(() => {
+    setOpen(false);
+  }, [location]);
+
+  return (
+    <Hidden mdUp>
+      <AppBar position="static" style={{ margin: 10 }}>
+        <Toolbar>
+          <IconButton edge="start" color="inherit" aria-label="menu" onClick={() => setOpen(true)}>
+            <MenuIcon />
+          </IconButton>
+          {links
+            .filter(link => link.path === location.pathname)
+            .map(link => (
+              <Tab key={link.path} label={link.label} component={NavLink} to={link.path} value={link.path} />
+            ))}
+        </Toolbar>
+      </AppBar>
+      <Drawer
+        className={classes.drawer}
+        onClose={() => setOpen(false)}
+        anchor="left"
+        open={open}
+        classes={{
+          paper: classes.drawerPaper
+        }}>
+        <div className={classes.drawerHeader}>
+          <IconButton onClick={() => setOpen(false)}>
+            {theme.direction === "ltr" ? <ChevronLeftIcon /> : <ChevronRightIcon />}
+          </IconButton>
+        </div>
+        <Divider />
+        <List>
+          {links.map(link => (
+            <ListItem
+              button
+              key={link.label}
+              selected={link.path === location.pathname}
+              component={NavLink}
+              to={link.path}>
+              <ListItemIcon>
+                <ChevronRightIcon />
+              </ListItemIcon>
+              <ListItemText primary={link.label} />
+            </ListItem>
+          ))}
+        </List>
+      </Drawer>
+    </Hidden>
+  );
+};
+
+export default MenuMobile;
